Combine request/response into single API file for simplicity
