+++ lib/cloudapp-export.rb
@@ -1,10 +1,7 @@
+require "cloudapp-export/api"
 require "cloudapp-export/cli"
 require "cloudapp-export/item"
-require "cloudapp-export/request"
-require "cloudapp-export/response"
 require "cloudapp-export/version"
 
-module Cloudapp
-  module Export
-  end
+module CloudappExport
 end
